Support Sphinx inline parameter syntax.
Sphinx allows types of parameters to be specified in the :param tag without
a separate :type line, such as ":param int a: desc", which results in a more
compact docstring.
This change adds support by extending the docstring parser to parse these lines.
Various test cases are included.

Note: My IDE removed a couple of trailing whitespaces in the edited files.