towards python 3 support
add a with_metaclass top level function which allows metaclass to work
syntactically in both python 2 and python 3.

don't test the python 3 only syntax test file with python 2.

---

The remaining failing tests are `contracts.interface.ContractSyntaxError`s.
It's unclear why these are failing, they appear to be stateful (as the
code works correctly outside of the context of the test suite!). Could this
be caching related?

*Note: Sometimes the tests do pass on python 3.3 and 3.4!!*